requirements.txt
Flask==3.0.0
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
WTForms
Flask-Login
Flask-Mail
Flask-Cors
psycopg2-binary
gunicorn
python-dotenv
requests
Jinja2
MarkupSafe
itsdangerous
Werkzeug
email-validator
Bootstrap-Flask
Flask-Admin
Flask-Uploads
Flask-Session
Flask-Bcrypt
Flask-RESTful
Flask-JWT-Extended
Pillow